+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## [1.1.0] ⁠– 2021-18-12
+
+### Feature
+
+- Burst-friendly complex number struct with basic utilities.
+- Polar decomposition, i.e. _A = U · P_.
+
 ## [1.0.0] ⁠– 2021-12-12
 
 ### Features